# kbot-final/pkg/sdk/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "kbot 企业级 AI Agent 教学平台 REST API，由 swaggo/swag 从 handler 注解生成。",
        "title": "kbot API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/kbs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "kbs"
                ],
                "summary": "列出知识库",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "kbs"
                ],
                "summary": "创建知识库",
                "parameters": [
                    {
                        "description": "知识库",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kb.CreateKBRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/kbs/{kb_id}/search": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "kbs"
                ],
                "summary": "KB 检索(bm25/vector/hybrid)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "KB ID",
                        "name": "kb_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "查询 + 模式 + top_k",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/prompts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "列出 Prompt",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "创建 Prompt(含 v1)",
                "parameters": [
                    {
                        "description": "Prompt",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prompt.CreatePromptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/prompts/{prompt_id}/promote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "晋升 Prompt 版本到某环境",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt ID",
                        "name": "prompt_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.GenerationConfig": {
            "type": "object",
            "properties": {
                "max_output_tokens": {
                    "type": "integer"
                },
                "seed": {
                    "type": "integer"
                },
                "stop": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "temperature": {
                    "type": "number"
                },
                "top_p": {
                    "type": "number"
                }
            }
        },
...
        "kb.CreateKBRequest": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "embedding_model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
        "prompt.CreatePromptRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "generation_config": {
                    "$ref": "#/definitions/domain.GenerationConfig"
                },
                "model_profile_version_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                },
                "variables_schema": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.SearchRequest": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "top_k": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}